pyproject.toml
[project]
dependencies = [
  "httpx>=0.28.1",
  "pendulum>=3.0.0",
  "pydantic>=2.10.6",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
  "rich>=13.9.4",
  "textual>=2.1.2",
  "textual-dev>=1.7.0",
]
description = "Add your description here"
name = "my-zt-tui"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "python-dotenv>=1.0.1",
]
[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib",
]
pythonpath = [
  "src",
]
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]
[tool.hatch.build.targets.wheel]
packages = ["src/my_zt"]
[project.scripts]
my_zt = "my_zt.entry_points:my_zt"
